New chatbot architecture LlmWithKoCodeTools
This chatbot uses one LLM with the Knowledge Object (KO) implementations
registered as function tools. We therefore guarantee that the
KO implementation code is run exactly as written in native Python.
We do not depend on LLMs to transcribe the code and we do not
depend on an interpreter model to run the code.

Performance on comprehensive_convo:
Runtime: 00:09:49
Cost: $0.74
Correctness:
Keyword: 0.68
ROUGE-L: 0.43
LLM:     0.78